lab-doc-agent/requirements.txt
anthropic>=0.40.0
notion-client>=2.2.1
gitpython>=3.1.41
python-dotenv>=1.0.1
